% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{eth.shp}
\alias{eth.shp}
\title{GADM 3.6 Ethiopia}
\format{
A sf object with GADM level 1 data for Ethiopia
}
\source{
https://gadm.org/download_country_v3.html
}
\usage{
eth.shp
}
\description{
Gadm level 1 data for Ethiopia
}
\keyword{datasets}
