% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmssamptrain-package.R
\docType{package}
\name{lsmssamptrain-package}
\alias{lsmssamptrain}
\alias{lsmssamptrain-package}
\title{lsmssamptrain: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Wild \email{Wild@worldbank.org}

}
\keyword{internal}
