% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population.hh}
\alias{population.hh}
\title{Synthetic population data for Ethiopia aggregated for the household level}
\format{
A tibble with synthetic population data for Ethiopia at the household level
}
\source{
generated with the synthpop package
}
\usage{
population.hh
}
\description{
Synthetic population data for Ethiopia
}
\keyword{datasets}
