% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{Synthetic population data for Ethiopia}
\format{
A tibble with synthetic population data for Ethiopia at the person level
}
\source{
generated with the synthpop package
}
\usage{
population
}
\description{
Synthetic population data for Ethiopia
}
\keyword{datasets}
